pyproject.toml
[project]
name = "weather-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.3.1",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
